Moved all the legacy code to the Deprecated namespace
