+++ Editor/Deprecated/TypeCollector.cs
@@ -1,10 +1,11 @@
-﻿namespace TypeReferences.Editor
+﻿namespace TypeReferences.Editor.Deprecated
 {
     using System;
     using System.Collections.Generic;
     using System.IO;
     using System.Linq;
     using System.Reflection;
+    using TypeReferences.Deprecated;
     using UnityEngine;
 
     /// <summary>
