+++ Editor/Deprecated/TypeDropDownDrawer.cs
@@ -1,9 +1,10 @@
-﻿namespace TypeReferences.Editor
+﻿namespace TypeReferences.Editor.Deprecated
 {
     using System;
     using System.Collections.Generic;
     using System.Linq;
     using System.Reflection;
+    using TypeReferences.Deprecated;
     using UnityEditor;
     using UnityEngine;
