+++ Runtime/Deprecated/Attributes/ClassExtendsAttribute.cs
@@ -1,4 +1,4 @@
-﻿namespace TypeReferences
+﻿namespace TypeReferences.Deprecated
 {
     using System;
