+++ Runtime/Deprecated/ClassTypeReference.cs.meta
